update:
Update CSS for theory-test-paper
