+++ Frontend/WebApplication/src/bigcomponents/MemberPage/components/molecules/theory-paper/theory-test.scss
@@ -1,0 +1,172 @@
+.theory-paper-container {
+    padding-top: 50px;
+
+    .theory-paper-title {
+        .paper-title {
+            text-transform: uppercase;
+            text-align: center;
+            text-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
+        }
+    }
+
+    .theory-paper-box {
+        margin-top: 100px;
+        width: 85%;
+        margin-left: auto;
+        margin-right: auto;
+
+        form {
+            display: grid;
+            position: relative;
+            grid-template-columns: 1fr 2fr;
+
+            table {
+                width: 100%;
+                border: 0.5mm solid black;
+                height: auto;
+
+                thead {
+                    .question-number-container {
+                        text-align: center;
+
+                        .countdown-timer {
+                            display: flex;
+                            align-items: center;
+                            padding: 30px 0;
+                            vertical-align: middle;
+                            justify-content: center;
+                            line-height: inherit;
+                            text-align: center;
+                            gap: 5px;
+                            font-size: 24px;
+
+                            .clock-icon {
+                                font-size: 36px;
+                                vertical-align: middle;
+                                color: #4292EB;
+                            }
+
+                            span {
+                                margin-left: 3px;
+                            }
+                        }
+
+                        .top-line {
+                            background-color: black;
+                            width: 100%;
+                            height: 0.5mm;
+                        }
+                    }
+                }
+
+                tbody {
+                    height: inherit;
+
+                    .questions-circle {
+                        height: inherit;
+                        padding: 20px 20px;
+
+                        ul {
+                            padding: 0 !important;
+                            margin: 0;
+                            display: grid;
+                            grid-template-columns: repeat(5, auto);
+                            /* Create 5 columns with equal width */
+                            grid-template-rows: repeat(7, auto);
+                            /* Create 7 rows with auto height */
+                            align-items: center;
+                            row-gap: 30px;
+                            justify-items: center;
+
+                            li {
+                                display: flex;
+                                justify-content: center;
+                                align-items: center;
+                                width: 40px;
+                                height: 40px;
+                                font-size: 16px;
+                                border: 1px solid black;
+                                border-radius: 50%;
+                                background: rgba(0, 162, 255, 0.1);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        .question-content-container {
+            width: 90%;
+            height: auto;
+            border: 0.5mm solid black;
+            justify-self: flex-end;
+            padding: 5px 5px;
+
+            .content-box {
+
+                .question-title {
+                    display: grid;
+                    grid-template-columns: 2fr 1fr;
+
+                    .question-num,
+                    .question-notice {
+                        display: inline-block;
+                    }
+
+                    .question-notice {
+                        font-style: italic;
+                        color: red;
+                        align-self: center;
+                        justify-self: flex-end;
+                    }
+
+                }
+
+                .question-content {
+                    font-weight: 600;
+                    font-size: 20px;
+                }
+            }
+
+            .answer-box {
+                display: grid;
+                grid-template-rows: repeat(3, 1fr);
+                gap: 40px;
+                padding: 10px 10px;
+
+                .answer {
+                    padding: 10px 10px;
+                    height: calc(350px / 3);
+                    width: 100%;
+                    border: 1px solid black;
+
+                    input {
+                        width: 15px;
+                        height: 15px;
+                        margin-right: 5px;
+                    }
+                }
+
+            }
+
+            .button-box {
+                padding: 10px 10px;
+                display: grid;
+                grid-template-columns: 1fr 1fr;
+
+                .previous-btn {
+                    width: 30%;
+                }
+
+                .next-btn {
+                    justify-self: flex-end;
+                }
+            }
+        }
+
+        .submit-btn {
+            margin-top: 20px;
+            text-align: center;
+        }
+    }
+}
